Added open_todo.py
I added a simple script to orgnize daily todo files. It creates a new
markdown file every day and opens it in your favourite editor. It was
developed to work with Typora on Windows using python3. To get it to work
on other systems you would probably only need to updated the last line of
the script.